add reservation to dashboard

+++ api/src/main/java/io/epf/onlyroom/entity/reservation/ReservationController.java
@@ -1,6 +1,7 @@
 package io.epf.onlyroom.entity.reservation;
 
 import io.epf.onlyroom.entity.room.Room;
+import io.epf.onlyroom.entity.room.RoomDAO;
 import io.epf.onlyroom.payload.response.MessageResponse;
 import io.epf.onlyroom.security.services.UserDetailsImpl;
 import org.springframework.http.HttpStatus;
@@ -18,19 +19,22 @@ import java.util.List;
 public class ReservationController {
 
     private final ReservationDAO reservationDAO;
+    private final RoomDAO roomDAO;
 
-    public ReservationController(ReservationDAO reservationDAO) {
+    public ReservationController(ReservationDAO reservationDAO, RoomDAO roomDAO) {
         this.reservationDAO = reservationDAO;
+        this.roomDAO = roomDAO;
     }
 
     @GetMapping()
     @PreAuthorize("hasRole('USER')")
-    public List<Reservation> getReservations() {
-        Iterable<Reservation> it = this.reservationDAO.findAll();
-        List<Reservation> reservations = new ArrayList<>();
-        it.forEach(reservations::add);
-
-        return reservations;
+    public List<ResaWithRoom> getReservations(Authentication authentication) {
+        List<Reservation> reservations = this.reservationDAO.findByUserId(((UserDetailsImpl)authentication.getPrincipal()).getId());
+        List<ResaWithRoom> resa = new ArrayList<>();
+        for (Reservation reservation : reservations) {
+            resa.add(new ResaWithRoom(reservation, reservation.getRoom()));
+        }
+        return resa;
     }
 
     @PostMapping()
@@ -56,8 +60,27 @@ public class ReservationController {
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('USER')")
-    public void deleteReservation(@PathVariable Long id) {
-        this.reservationDAO.deleteById(id);
+    public ResponseEntity<MessageResponse> deleteReservation(@PathVariable Long id, Authentication authentication) {
+        if(this.reservationDAO.findById(id).get().getUser().getId() != ((UserDetailsImpl)authentication.getPrincipal()).getId()) {
+            return ResponseEntity
+                    .badRequest()
+                    .body(new MessageResponse("Error: user not authorized!"));
+        } else {
+            this.reservationDAO.deleteById(id);
+            return new ResponseEntity("Reservation deleted", HttpStatus.OK);
+        }
+
     }
 
+}
+
+class ResaWithRoom<Reservation, Room> {
+    public final Reservation reservation;
+    public final Room room;
+
+
+    ResaWithRoom(Reservation reservation, Room room) {
+        this.reservation = reservation;
+        this.room = room;
+    }
 }
